# openapi.yaml
openapi: 3.1.0
info:
  title: Persistor API
  description: Knowledge graph and vector memory service with tenant isolation, encryption at rest, and real-time notifications.
  version: 0.7.0
  license:
    name: AGPL-3.0
servers:
  - url: http://localhost:3030/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key mapped to a single tenant. SHA-256 hashed before storage.
  schemas:
    Node:
      type: object
      properties:
        id:
          type: string
          maxLength: 255
          description: Unique identifier. Auto-generated UUID if omitted on create.
        type:
          type: string
          maxLength: 100
          description: "Category: person, project, concept, event, place, etc."
        label:
          type: string
          maxLength: 10000
          description: Human-readable name. Used for full-text search.
        properties:
          type: object
          description: Arbitrary metadata. Encrypted at rest (AES-256-GCM).
        salience_score:
          type: number
          format: double
        access_count:
          type: integer
        last_accessed:
          type:
            - string
            - "null"
          format: date-time
        user_boosted:
          type: boolean
        superseded_by:
          type:
            - string
            - "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NodeCreate:
      type: object
      required: [type, label]
      properties:
        id:
          type: string
          maxLength: 255
        type:
          type: string
          maxLength: 100
        label:
          type: string
          maxLength: 10000
        properties:
          type: object
    NodeUpdate:
      type: object
      properties:
        type:
          type: string
          maxLength: 100
        label:
          type: string
          maxLength: 10000
        properties:
          type: object
    NodeMigrate:
      type: object
      required: [new_id]
      properties:
        new_id:
          type: string
          maxLength: 255
...
    PatchPropertiesRequest:
      type: object
      required: [properties]
      description: >
        Partially update properties with merge semantics. Keys with non-null
        values are added or updated; keys with explicit null values are removed.
      properties:
        properties:
          type: object
          description: Key-value pairs to merge. Null values delete the key.
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
paths:
...
  /nodes:
    get:
      summary: List nodes
      operationId: listNodes
      tags: [Nodes]
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: min_salience
          in: query
          schema:
            type: number
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 1000
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            maximum: 100000
      responses:
        "200":
          description: Paginated node list
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Node"
                  has_more:
                    type: boolean

    post:
      summary: Create a node
      operationId: createNode
      tags: [Nodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeCreate"
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "409":
          description: Node ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /nodes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get a node
      operationId: getNode
      tags: [Nodes]
      responses:
        "200":
          description: Node found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update a node
      operationId: updateNode
      tags: [Nodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeUpdate"
      responses:
        "200":
          description: Node updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a node
      description: Cascades — also deletes all connected edges.
      operationId: deleteNode
      tags: [Nodes]
      responses:
        "200":
          description: Node deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /nodes/{id}/properties:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    patch:
      summary: Partially update node properties
      description: >
        Merges the supplied properties into the existing node properties.
        Keys with non-null values are added or updated; keys with explicit
        null values are removed. Other existing properties are left intact.
      operationId: patchNodeProperties
      tags: [Nodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchPropertiesRequest"
      responses:
        "200":
          description: Node with updated properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /nodes/{id}/migrate:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Migrate a node to a new ID
      operationId: migrateNode
      tags: [Nodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeMigrate"
      responses:
        "200":
          description: Node migrated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
...